+++ packages/plugin-i18n/src/shared/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './use-i18n';
+export * from './use-static-translation';